Remove hidden coupling from SQLite contract checks
The unique-error check depended on a marker written by the pool check,
so changing that marker reported a failure in the wrong observation.
Duplicate the current row value so each check can change independently.

Keep driver-independent validation outside the parity observations and
remove duplicate busy and mattn checks to avoid redundant failures.
